Move to local components namespace
