+++ osu.Game/Overlays/Profile/ProfileHeader.cs
@@ -16,6 +16,7 @@ using osu.Game.Graphics;
 using osu.Game.Graphics.Containers;
 using osu.Game.Graphics.Sprites;
 using osu.Game.Graphics.UserInterface;
+using osu.Game.Overlays.Profile.Components;
 using osu.Game.Overlays.Profile.Header;
 using osu.Game.Users;
 
@@ -469,44 +470,6 @@ namespace osu.Game.Overlays.Profile
             }
 
             infoTextRight.NewLine();
-        }
-
-        private class GradeBadge : Container
-        {
-            private const float width = 50;
-            private readonly string grade;
-            private readonly Sprite badge;
-            private readonly SpriteText numberText;
-
-            public int DisplayCount
-            {
-                set { numberText.Text = value.ToString(@"#,0"); }
-            }
-
-            public GradeBadge(string grade)
-            {
-                this.grade = grade;
-                Width = width;
-                Height = 41;
-                Add(badge = new Sprite
-                {
-                    Width = width,
-                    Height = 26
-                });
-                Add(numberText = new OsuSpriteText
-                {
-                    Anchor = Anchor.BottomCentre,
-                    Origin = Anchor.BottomCentre,
-                    TextSize = 14,
-                    Font = @"Exo2.0-Bold"
-                });
-            }
-
-            [BackgroundDependencyLoader]
-            private void load(TextureStore textures)
-            {
-                badge.Texture = textures.Get($"Grades/{grade}");
-            }
         }
     }
 }
